Topic7MCP/requirements.txt
requests
python-dotenv
openai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
gradio>=4.0.0
